requirements.txt
pyyaml>=6.0.1
pytest>=8.0.0
pytest-cov>=4.1.0
python-dateutil
python-Levenshtein>=0.25.0
colorama>=0.4.6
